+++ .github/.env
@@ -1,0 +1,5 @@
+FLOWTHRU_DB_HOST=localhost
+FLOWTHRU_DB_PORT=5432
+FLOWTHRU_DB_USER=user
+FLOWTHRU_DB_PASSWORD=password
+FLOWTHRU_DB_NAME=test
