+++ internal/repository/pg/repository_test.go
@@ -1,4 +1,4 @@
-package pgsql
+package pg
 
 import (
     "testing"
@@ -6,23 +6,20 @@ import (
     "fmt"
 )
 
-
-var DBTestCreds = fmt.Sprintf(
-    "host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
-    os.Getenv("FLOWTHRU_TEST_DB_HOST"),
-    os.Getenv("FLOWTHRU_TEST_DB_PORT"),
-    os.Getenv("FLOWTHRU_TEST_DB_USER"),
-    os.Getenv("FLOWTHRU_TEST_DB_PASSWORD"),
-    os.Getenv("FLOWTHRU_TEST_DB_NAME"),
-)
-
 // TestNewDB tests if a connections is correctly established.
 // Additionally checks if a disconnection is completed without errors
 func TestNewDB(t *testing.T) {
-    db, err := NewDB(DBTestCreds)
+    db, err := NewDB(
+            "host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
+            os.Getenv("FLOWTHRU_DB_HOST"),
+            os.Getenv("FLOWTHRU_DB_PORT"),
+            os.Getenv("FLOWTHRU_DB_USER"),
+            os.Getenv("FLOWTHRU_DB_PASSWORD"),
+            os.Getenv("FLOWTHRU_DB_NAME"))))
     if err != nil {
         t.Fatalf("Error when connecting to database: %q", err.Error())
     }
+
     err = db.Close()
     if err != nil {
         t.Fatalf("Error when closing connection to database: %q", err.Error())
