--- server/.dockerignore
@@ -1,3 +1,0 @@
-
-*.log
-bin/
