--- server/Dockerfile
@@ -1,10 +1,0 @@
-
-FROM golang:1.12.7-stretch
-WORKDIR /go/src/server
-
-COPY . .
-RUN GO111MODULE=on go build -o /bin/app cmd/flowthru-server/main.go
-
-EXPOSE 5000
-CMD ["/bin/app"]
-
