--- server/Dockerrun.aws.json
@@ -1,9 +1,0 @@
-{
-  "AWSEBDockerrunVersion": "1",
-  "Ports": [
-    {
-      "ContainerPort": "5000"
-    }
-  ],
-  "Logging": "/var/log/nginx"
-}
