--- server/go.mod
@@ -1,13 +1,0 @@
-module server
-
-go 1.12
-
-require (
-	github.com/auth0/go-jwt-middleware v0.0.0-20190805220309-36081240882b
-	github.com/codegangsta/negroni v1.0.0
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/gorilla/mux v1.7.3
-	github.com/gorilla/sessions v1.2.0 // indirect
-	github.com/julienschmidt/httprouter v1.2.0
-	github.com/lib/pq v1.2.0
-)
